refactor: 서버 전용 api 따로 분리
